Split classes into files. Reformating with PEP
