+++ poker/gui/plots/bar_plotter.py
@@ -1,0 +1,77 @@
+from weakref import proxy
+
+import numpy as np
+from matplotlib.backends.backend_qt5agg import (
+    FigureCanvasQTAgg as FigureCanvas)
+from matplotlib.figure import Figure
+
+
+class BarPlotter(FigureCanvas):
+    def __init__(self, ui, p):
+        self.p = p
+        self.ui = proxy(ui)
+        self.fig = Figure(dpi=50)
+        self.axes = self.fig.add_subplot(111)  # create an axis
+        super(BarPlotter, self).__init__(self.fig)
+        # self.drawfigure()
+        self.ui.vLayout2.insertWidget(1, self)
+
+    def drawfigure(self, l, strategy):
+        self.axes = self.fig.add_subplot(111)  # create an axis
+
+        data = l.get_stacked_bar_data('Template', strategy, 'stackedBar')
+
+        N = 11
+        Bluff = data[0]
+        BP = data[1]
+        BHP = data[2]
+        Bet = data[3]
+        Call = data[4]
+        Check = data[5]
+        Fold = data[6]
+        ind = np.arange(N)  # the x locations for the groups
+        width = 1  # the width of the bars: can also be len(x) sequence
+
+        self.p0 = self.axes.bar(ind, Bluff, width, color='y')
+        self.p1 = self.axes.bar(ind, BP, width, color='k', bottom=Bluff)
+        self.p2 = self.axes.bar(ind, BHP, width, color='b', bottom=[sum(x) for x in zip(Bluff, BP)])
+        self.p3 = self.axes.bar(ind, Bet, width, color='c', bottom=[sum(x) for x in zip(Bluff, BP, BHP)])
+        self.p4 = self.axes.bar(ind, Call, width, color='g', bottom=[sum(x) for x in zip(Bluff, BP, BHP, Bet)])
+        self.p5 = self.axes.bar(ind, Check, width, color='w', bottom=[sum(x) for x in zip(Bluff, BP, BHP, Bet, Call)])
+        self.p6 = self.axes.bar(ind, Fold, width, color='r',
+                                bottom=[sum(x) for x in zip(Bluff, BP, BHP, Bet, Call, Check)])
+
+        self.axes.set_ylabel('Profitability')
+        self.axes.set_title('FinalFundsChange ABS')
+        self.axes.set_xlabel(['PF Win', 'Loss', '', 'F Win', 'Loss', '', 'T Win', 'Loss', '', 'R Win', 'Loss'])
+        self.axes.legend((self.p0[0], self.p1[0], self.p2[0], self.p3[0], self.p4[0], self.p5[0], self.p6[0]),
+                         ('Bluff/Decept.', 'BetPot', 'BetHfPot', 'Bet/Bet+', 'Call', 'Check', 'Fold'),
+                         labelspacing=0.03,
+                         prop={'size': 12})
+        maxh = float(self.p.selected_strategy['bigBlind']) * 20
+        i = 0
+        for rect0, rect1, rect2, rect3, rect4, rect5, rect6 in zip(self.p0.patches, self.p1.patches,
+                                                                   self.p2.patches,
+                                                                   self.p3.patches, self.p4.patches,
+                                                                   self.p5.patches, self.p6.patches):
+            g = list(zip(data[0], data[1], data[2], data[3], data[4], data[5], data[6]))
+            height = g[i]
+            i += 1
+            rect0.set_height(height[0])
+            rect1.set_y(height[0])
+            rect1.set_height(height[1])
+            rect2.set_y(height[0] + height[1])
+            rect2.set_height(height[2])
+            rect3.set_y(height[0] + height[1] + height[2])
+            rect3.set_height(height[3])
+            rect4.set_y(height[0] + height[1] + height[2] + height[3])
+            rect4.set_height(height[4])
+            rect5.set_y(height[0] + height[1] + height[2] + height[3] + height[4])
+            rect5.set_height(height[5])
+            rect6.set_y(height[0] + height[1] + height[2] + height[3] + height[4] + height[5])
+            rect6.set_height(height[6])
+            maxh = max(height[0] + height[1] + height[2] + height[3] + height[4] + height[5] + height[6], maxh)
+
+        self.axes.set_ylim((0, maxh))
+
+        self.draw()
