+++ poker/gui/ui/gui_qt_ui_analyser.ui
