+++ poker/gui/ui/gui_qt_ui_genetic_algorithm.ui
