+++ poker/gui/ui/gui_qt_ui_strategy_manager.ui
