+++ poker/gui/ui/ui_qt_genetic_algorithm.ui
